feat!: delete wirePFB
