+++ x/blob/client/cli/tx.go
@@ -21,5 +21,7 @@ func GetTxCmd() *cobra.Command {
 		RunE:                       client.ValidateCmd,
 	}
 
+	cmd.AddCommand(CmdWirePayForBlob())
+
 	return cmd
 }
