+++ x/blob/types/codec.go
@@ -12,15 +12,10 @@ import (
 var ModuleCdc = codec.NewProtoCodec(codectypes.NewInterfaceRegistry())
 
 func RegisterCodec(cdc *codec.LegacyAmino) {
-	cdc.RegisterConcrete(&MsgWirePayForBlob{}, URLMsgWirePayForBlob, nil)
 	cdc.RegisterConcrete(&MsgPayForBlob{}, URLMsgPayForBlob, nil)
 }
 
 func RegisterInterfaces(registry codectypes.InterfaceRegistry) {
-	registry.RegisterImplementations((*sdk.Msg)(nil),
-		&MsgWirePayForBlob{},
-	)
-
 	registry.RegisterImplementations((*sdk.Msg)(nil),
 		&MsgPayForBlob{},
 	)
